Created CreateUserTests and GetAllUsersTests class and segregated the tests
